Revert "fix: exclude env key 'PROJECT_ENV' and evaluate referenced variables (#4275)"
This reverts commit 58c3b9b538218d59dd7dc985c8c9b6a4295ed639.
