+++ crates/pixi_core/src/task/executable_task.rs
@@ -1,6 +1,6 @@
 use std::{
     borrow::Cow,
-    collections::HashMap,
+    collections::{HashMap, HashSet},
     ffi::OsString,
     fmt::{Display, Formatter},
     path::PathBuf,
@@ -451,7 +451,22 @@ fn get_export_specific_task_env(task: &Task, command_env: IndexMap<String, Strin
     }
 
     // Define keys that should not be overridden
-    let override_excluded_keys = consts::get_override_excluded_keys();
+    let override_excluded_keys: HashSet<&str> = [
+        "PIXI_PROJECT_ROOT",
+        "PIXI_PROJECT_NAME",
+        "PIXI_PROJECT_VERSION",
+        "PIXI_PROMPT",
+        "PIXI_ENVIRONMENT_NAME",
+        "PIXI_ENVIRONMENT_PLATFORMS",
+        "CONDA_PREFIX",
+        "CONDA_DEFAULT_ENV",
+        "PATH",
+        "INIT_CWD",
+        "PWD",
+    ]
+    .iter()
+    .cloned()
+    .collect();
 
     // Create environment map struct
     let env_map = EnvMap::new(command_env.clone(), task.env());
@@ -461,7 +476,7 @@ fn get_export_specific_task_env(task: &Task, command_env: IndexMap<String, Strin
         .expect("Task environment should exist at this point");
 
     // Determine export strategy
-    let mut export_merged = if task_env.keys().all(|k| !command_env.contains_key(k)) {
+    let export_merged = if task_env.keys().all(|k| !command_env.contains_key(k)) {
         // If task.env() and command_env don't have duplicated keys, simply export task.env().
         task_env.clone()
     } else {
@@ -469,7 +484,6 @@ fn get_export_specific_task_env(task: &Task, command_env: IndexMap<String, Strin
         env_map.merge_by_priority()
     };
 
-    Environment::resolve_variable_references(&mut export_merged);
     // Build export string
     // Put all merged environment variables to export.
     // Only the keys that are in "task_specific_envs" map would be exported.
